refactor(datafusion): merge IcebergPartitionedTableProvider into IcebergTableProvider
IcebergPartitionedTableProvider and IcebergPartitionedScan were introduced
to enable parallel file scanning by bucketing FileScanTasks across DataFusion
partitions. However, maintaining two TableProvider implementations is
redundant: the new provider is strictly more capable, and its degenerate case
(target_partitions=1) reproduces the old single-partition behavior exactly.

This commit folds the partitioned provider into IcebergTableProvider and
the partitioned scan into IcebergTableScan, eliminating the parallel types.

Changes:
- IcebergTableProvider::scan() now eagerly calls plan_files() and distributes
FileScanTasks into buckets using the same identity-hash strategy
(REPARTITION_RANDOM_STATE + create_hashes) that was in
IcebergPartitionedTableProvider, enabling Partitioning::Hash declarations
that align with DataFusion's RepartitionExec.
- IcebergTableScan gains a new_with_tasks() constructor that accepts
pre-planned buckets and a caller-supplied Partitioning. execute(i) streams
the tasks in buckets[i] via TableScan::to_arrow_with_tasks, rebuilding
the TableScan per-partition to avoid serializing PlanContext Arc-shared
caches across workers.
- The original new() constructor and the to_arrow() lazy path are kept
unchanged for IcebergStaticTableProvider, which does not pre-plan tasks.
- Limit slicing (try_filter_map truncation) from the old IcebergTableScan
is preserved in both execution paths.
- Bucketing helpers (IdentityCol, compute_identity_cols, bucket_tasks,
identity_hash, fallback_hash, literal_to_array, is_supported_dtype) are
moved verbatim into a new private table/bucketing.rs module.
- Unit tests from partitioned.rs are migrated to table/mod.rs and updated
to use IcebergTableProvider and IcebergTableScan.
- integration_datafusion_test.rs: fix test_provider_plan_stream_schema to
call execute(0) instead of execute(1). The old call worked only because
the previous IcebergTableScan silently ignored the partition index.

(cherry picked from commit d2e5e0412c1e7e17f85f9fa549af1544d07eaae1)
