+++ crates/integrations/datafusion/src/lib.rs
@@ -24,8 +24,6 @@ pub use error::*;
 pub mod physical_plan;
 mod schema;
 pub mod table;
-pub use physical_plan::IcebergPartitionedScan;
-pub use table::partitioned::IcebergPartitionedTableProvider;
 pub use table::table_provider_factory::IcebergTableProviderFactory;
 pub use table::*;
