+++ crates/integrations/datafusion/src/physical_plan/mod.rs
@@ -18,7 +18,6 @@
 pub(crate) mod commit;
 pub(crate) mod expr_to_predicate;
 pub(crate) mod metadata_scan;
-pub(crate) mod partitioned_scan;
 pub(crate) mod project;
 pub(crate) mod repartition;
 pub(crate) mod scan;
@@ -28,6 +27,5 @@ pub(crate) mod write;
 pub(crate) const DATA_FILES_COL_NAME: &str = "data_files";
 
 pub use expr_to_predicate::convert_filters_to_predicate;
-pub use partitioned_scan::IcebergPartitionedScan;
 pub use project::project_with_partition;
 pub use scan::IcebergTableScan;
